test: move form field tests into separate files
